sentiment-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
kafka-python-ng==2.2.2
transformers==4.38.1
torch==2.2.0
numpy<2,>=1.26.4
vaderSentiment==3.3.2
pydantic==2.6.1
pydantic-settings==2.1.0
prometheus-client==0.20.0
